LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-history-back-after-fragment-expected.txt


FAIL history.back() fires the navigate event and sets hashChange when reversing a fragment navigation assert_equals: expected (object) null but got (undefined) undefined
